// src/server/schemas/pokemon.ts
export const pokemonAPI = z.object({
// ...
  sprites: z.object({
    back_default: z.string(),
    back_female: z.null(),
    back_shiny: z.string(),
    back_shiny_female: z.null(),
    front_default: z.string(),
    front_female: z.null(),
    front_shiny: z.string(),
    front_shiny_female: z.null(),
    other: z.object({
      dream_world: z.object({
        front_default: z.string(),
        front_female: z.null()
      }),
      home: z.object({
        front_default: z.string(),
        front_female: z.null(),
        front_shiny: z.string(),
        front_shiny_female: z.null()
      }),
      "official-artwork": z.object({
        front_default: z.string(),
        front_shiny: z.string()
      })
    }),
// ...
  }),
  stats: z.array(
    z.object({
      base_stat: z.number(),
      effort: z.number(),
      stat: z.object({ name: z.string(), url: z.string() })
    })
  ),
  types: z.array(
    z.object({
      slot: z.number(),
      type: z.object({ name: z.string(), url: z.string() })
    })
  ),
  weight: z.number()
})